net: procfs: reduce rcu lock critical section
The current /proc/net/dev sequence file implementation, which shows
statistics of the interfaces from the current netns, holds rcu_read_lock()
throughout the whole process, from dev seq start up to dev seq stop.

This is an issue because there is an effort to ensure that dev_get_stats
is able to sleep, and when the caller holds rcu_read_lock(), the
resulting context is atomic.

The RCU read-side critical section is needed to guarantee that the
device chain is not modified while the dev sequence file is walking
through it and handling the netdev at the same time.

To minimize this critical section and drastically reduce the time rcu lock
is being held, all we need is to grab the rcu lock only for the brief
moment where we are looking for the next netdev to handle. If found,
dev_hold it to guarantee it kept alive while accessed later in seq show
callback and release the rcu lock immediately.

The current netdev being handled will be released using dev_put when the
seq next callback is called or dev seq stop is called.

Signed-off-by: Saeed Mahameed <saeed@kernel.org>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>