net: phylink: re-implement interface configuration with PCS
With PCS support, how we implement interface reconfiguration is not up
to the job; we end up reconfiguring the PCS for an interface change
while the link could potentially be up.  In order to solve this, add
two additional MAC methods for interface configuration, one to prepare
for the change, and one to finish the change.

This allows mvneta and mvpp2 to shutdown what they require prior to the
MAC and PCS configuration calls, and then restart as appropriate.

This impacts ksettings_set(), which now needs to identify whether the
change is a minor tweak to the advertisement masks or whether the
interface mode has changed, and call the appropriate function for that
update.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>