docs: networking: timestamping: replace tabs with spaces in code blocks
Reading the document in vim is currently not a pleasant experience. Its
rst syntax highlighting is confused by the "*/" sequences which it's not
interpreting as part of the code blocks for some reason.

Replace the tabs with spaces, so that syntax highlighters (at least the
one in vim) have a better idea where code blocks start and where they
end.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>