docs: networking: timestamping: add one more known issue
Document the fact that Ethernet PHY timestamping has a fundamentally
flawed corner case (which in fact hits the majority of networking
drivers): a PHY for which its host MAC driver doesn't forward the
phy_mii_ioctl for timestamping is still going to be presented to user
space as functional.

Fixing this inconsistency would require moving the phy_has_tsinfo()
check inside all MAC drivers which are capable of PHY timestamping, to
be in harmony with the existing design for phy_has_hwtstamp() checks.
Instead of doing that, document the preferable solution which is that
offending MAC drivers be fixed instead.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>