net: dsa: felix: setup MMIO filtering rules for PTP when using tag_8021q
Because we do want to support PTP even with tag_8021q, we need to use
the CPU port module for that. And because we can't use NPI mode, then we
need to poll the MMIO registers to retrieve TX and RX timestamps.

The only traffic for which we need timestamps is PTP traffic, so add a
redirection rule to the CPU port module for this.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>