net: dsa: felix: add a Kconfig-based workaround for the lack of an XTR IRQ
Since the tag_8021q tagger is software-defined, it has no means by
itself for retrieving PTP timestamps of packets. If PTP must be
supported by a user of tag_8021q, hardware-specific means need to be
deployed.

In the case of Ocelot switches, those hardware-specific means involve
the Injection/Extraction functionality. The RX timestamp is present in
the Extraction Frame Header.

Sadly, on NXP LS1028A, the Felix switch was integrated into the SoC
without wiring the extraction IRQ line to the ARM GIC. So, if we want to
be notified of any PTP packets received on the CPU port module, we have
a problem.

There is a possible workaround, which is to use the Ethernet CPU port as
a notification channel that packets are available on the CPU port module
as well.

Create a build-time kernel config option that is selected by the Felix
driver. We want to do this such that the workaround is minimally
invasive for the switches that don't require this workaround. The goal
is for the Felix-specific code to be completely compiled out at build
time.

This config option will be used incrementally in future patches.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>