net: mscc: ocelot: register devlink ports
Add devlink integration into the mscc_ocelot switchdev driver. Only the
probed interfaces are registered with devlink, because for convenience,
struct devlink_port was included into struct ocelot_port_private, which
is only initialized for the ports that are used.

Note that the felix DSA driver is already integrated with devlink by
default, since that is a thing that the DSA core takes care of.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>