net: dsa: sja1105: make devlink property best_effort_vlan_filtering true by default
For users, learning that the driver does not support traffic termination
from the network stack when the ports are under a bridge with
vlan_filtering=1 generates a 'wtf' moment. They need to go to the
documentation and find about the existence of this property, then maybe
install devlink and set it to true.

Having best_effort_vlan_filtering enabled by default delays that 'wtf'
moment (maybe up to the point that it never even happens). The user
doesn't need to care that the driver supports addressing the ports
individually by retagging VLAN IDs until he/she needs to use more than
32 VLAN IDs (since there can be at most 32 retagging rules).  Only then
do they need to think whether they need the full VLAN table, at the
expense of no individual port addressing, or not.

But the last time I heard of a user need more than 32 VLANs was... never.
So we should be able to safely change the default here, and make this
driver act just a little bit more sanely out of the box.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>